(Assets e index) Refinei o projeto pra colocar no codespaces

+++ corridas/delete.php
@@ -6,9 +6,10 @@ include '../config/conexao.php';
 if ($_SERVER['REQUEST_METHOD'] === 'GET') {
     $id = $_GET['id'];
 
-    $sql = "DELETE FROM corridas WHERE id = :id";
+    $sql = "DELETE FROM corridas WHERE id = :id AND user_id = :user_id";
     $stmt = $pdo->prepare($sql);
     $stmt->bindValue(':id', $id, PDO::PARAM_INT);
+    $stmt->bindValue(':user_id', $_SESSION['id'], PDO::PARAM_INT);
 
     if ($stmt->execute()) {
         header("Location: corridas.php");
